// modules/generatedCode/TwelveDataQuote.ts
export interface TwelveDataQuote {
    symbol:         string;
    name:           string;
    exchange:       string;
    mic_code:       string;
    currency:       string;
    datetime:       Date;
    timestamp:      number;
    open:           string;
    high:           string;
    low:            string;
    close:          string;
    volume:         string;
    previous_close: string;
    change:         string;
    percent_change: string;
    average_volume: string;
    is_market_open: boolean;
    fifty_two_week: FiftyTwoWeek;
}

export interface FiftyTwoWeek {
    low:                 string;
    high:                string;
    low_change:          string;
    high_change:         string;
    low_change_percent:  string;
    high_change_percent: string;
    range:               string;
}

// Converts JSON strings to/from your types
// and asserts the results of JSON.parse at runtime
export class ConvertQuote {
    public static toTwelveDataQuote(json: string): TwelveDataQuote {
        return cast(JSON.parse(json), r("TwelveDataQuote"));
    }

    public static twelveDataQuoteToJson(value: TwelveDataQuote): string {
        return JSON.stringify(uncast(value, r("TwelveDataQuote")), null, 2);
    }
}