// modules/generatedCode/TwelveDataError.ts
export interface TwelveDataError {
    code:    number;
    message: string;
    status:  string;
    meta:    Meta;
}

export interface Meta {
    symbol:   string;
    interval: string;
    exchange: string;
}

// Converts JSON strings to/from your types
// and asserts the results of JSON.parse at runtime
export class ConvertError {
    public static toTwelveDataError(json: string): TwelveDataError {
        return cast(JSON.parse(json), r("TwelveDataError"));
    }

    public static twelveDataErrorToJson(value: TwelveDataError): string {
        return JSON.stringify(uncast(value, r("TwelveDataError")), null, 2);
    }
}