// modules/generatedCode/TwelveDataQuoteExtendedHours.ts
export interface TwelveDataQuoteExtendedHours {
    symbol:                  string;
    name:                    string;
    exchange:                string;
    mic_code:                string;
    currency:                string;
    datetime:                Date;
    timestamp:               number;
    open:                    string;
    high:                    string;
    low:                     string;
    close:                   string;
    volume:                  string;
    previous_close:          string;
    change:                  string;
    percent_change:          string;
    average_volume:          string;
    rolling_1d_change:       string;
    rolling_7d_change:       string;
    rolling_period_change:   string;
    is_market_open:          boolean;
    fifty_two_week:          FiftyTwoWeek;
    extended_change:         string;
    extended_percent_change: string;
    extended_price:          string;
    extended_timestamp:      number;
}

export interface FiftyTwoWeek {
    low:                 string;
    high:                string;
    low_change:          string;
    high_change:         string;
    low_change_percent:  string;
    high_change_percent: string;
    range:               string;
}

// Converts JSON strings to/from your types
// and asserts the results of JSON.parse at runtime
export class ConvertQuoteExtendedHours {
    public static toTwelveDataQuoteExtendedHours(json: string): TwelveDataQuoteExtendedHours {
        return cast(JSON.parse(json), r("TwelveDataQuoteExtendedHours"));
    }

    public static TwelveDataQuoteExtendedHoursToJson(value: TwelveDataQuoteExtendedHours): string {
        return JSON.stringify(uncast(value, r("TwelveDataQuoteExtendedHours")), null, 2);
    }
}